requirements/base.txt
Django==1.7.4
psycopg2==2.6
djangorestframework==3.0.5
django-filter==0.9.2
django-bootstrap3==5.1.1
django-cors-headers==1.0.0
Pillow==2.7.0
